+++ next.config.mjs
@@ -1,12 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-  webpack: (config) => {
-    config.module.rules.push({
-      test: /\.css$/,
-      use: ['style-loader', 'css-loader', 'postcss-loader']
-    });
-    return config;
-  }
-};
+const nextConfig = {};
 
 export default nextConfig;
